Renames libs to bitcoin since only bitcoin related stuff was there.
