+++ tool/bitcoin/bitcoinBitcoindData.js
