+++ tool/bitcoin/bitcoinCoinbaseDataFromBitcoind.js
