+++ tool/getInformationReadyForRegisterBtcTransaction.js
@@ -8,8 +8,8 @@ const pmtBuilder = require("../index");
 const bitcoin = require("bitcoinjs-lib");
 const { createMempoolBitcoinClients } = require("./mempool-api-client");
 const { fetchBlockWtxidsWithTargetWtxid } = require("./pmt-builder-utils");
-const { getBitcoinTransactionDataForPmt } = require("./lib/bitcoinTransactionDataForPmt");
-const { bitcoindRpc } = require("./lib/bitcoindRpc");
+const { getBitcoinTransactionDataForPmt } = require("./bitcoin/bitcoinTransactionDataForPmt");
+const { bitcoindRpc } = require("./bitcoin/bitcoindRpc");
 
 const MEMPOOL_NETWORKS = new Set(["mainnet", "testnet"]);
